requirements.txt
certifi==2020.12.5
chardet==4.0.0
happybase==1.2.0
idna==2.10
ply==3.11
pycryptodome==3.10.1
requests==2.25.1
six==1.15.0
thriftpy2==0.4.14
urllib3==1.26.4

boto3~=1.17.69